feat: create DB

+++ app/src/main/java/com/bogdanmurzin/mypersonalwallet/data/transaction_recycer_items/RecyclerMultiTypeItem.kt
@@ -3,6 +3,6 @@ package com.bogdanmurzin.mypersonalwallet.data.transaction_recycer_items
 import java.math.BigDecimal
 
 interface RecyclerMultiTypeItem {
-    val transactionAmount: BigDecimal
+    val transactionAmount: Float
     val type: Int
 }
